\name{generateQuadraticLinearSpiral}
\alias{generateQuadraticLinearSpiral}
\title{
generate quadratic connected to a linear spiral (QL)
}
\description{generate Quadratic connected to Linear spiral (QL) with given change location.
}
\usage{
generateQuadraticLinearSpiral(nstart = 200, change = 100, k1 = 40, k2 = 50, q = 0, 
	dist.zero = 2, noise = 0.001, theta.start = 1e-04, theta.end = 1.25, 
        equal = FALSE, equal.n = 200, move = TRUE, move.factor = 1000, 
         plot.it = TRUE, show.change = TRUE)
}
\arguments{
  \item{nstart}{
number of outline points to generate
}
  \item{change}{
location of change (tau)
}
  \item{k1}{
first expansion angle
}
  \item{k2}{
second expansion angle
}
  \item{q}{
quadratic factor
}
  \item{dist.zero}{
distance between axis and first outline coordinate (angle = 0)
}
  \item{noise}{
noise level ( standard deviation in coordinates)
}
  \item{theta.start}{
initial angle at start of outline ( a small number)
}
  \item{theta.end}{
final angle at end of outline 
}
  \item{equal}{
create equally spaced outline coordinates
}
  \item{equal.n}{
number o equally spaced points
}
  \item{move}{
move the outline
}
  \item{move.factor}{
how much to move or translate the outline
}
  \item{plot.it}{
plot the outline or not
}
  \item{show.change}{
location the change 
}
}
\details{
no details yet
}
\value{
 data frame with spiral coordinates as x, y, logr, angle, and arc length
}
\references{
 no reference yet
}
\author{
A.E. Aldridge
}
\note{
not generated by equal arc lengths
}
\seealso{
generateAnyLinearSpirals in logspiral package
}
\examples{
\dontrun{
 z <- generateQuadraticLinearSpiral(k2=0,q= -0.5, equal=F )
xs <- fitQuadraticLinear( x=z$x, y=z$y, m=100, axis.start=c(990,990))
xs$parameters1
xs$parameters2
}
}
\keyword{ CurvedSpiral } 
