\name{generateQuadraticSpirals}
\alias{generateQuadraticSpirals}
\title{
generates TWO quadratic spirals (Q2) with given change location
}
\description{
generate TWO connected quadratic spirals with the given change location
}
\usage{
generateQuadraticSpirals(nstart = 200, change = 100, k1 = 40, k2 = 50, q1 = 0.03, 
             dist.zero = 2, noise = 0.001, theta.start = 1e-04, theta.end = 1.25, 
              equal = TRUE, equal.n = 200, move = TRUE, move.factor = 1000, 
              plot.it = TRUE, show.change = TRUE)
}
\arguments{
   \item{nstart}{
number of outline points to generate
}
  \item{change}{
location of change (tau)
}
  \item{k1}{
first expansion angle
}
  \item{k2}{
second expansion angle
}
  \item{q1}{
first quadratic term ( second is calculated using continuity)
}
  \item{dist.zero}{
distance between axis and first outline coordinate (angle = 0)
}
  \item{noise}{
noise level ( standard deviation in coordinates)
}
  \item{theta.start}{
initial angle at start of outline ( a small number)
}
  \item{theta.end}{
final angle at end of outline 
}
  \item{equal}{
create equally spaced outline coordinates
}
  \item{equal.n}{
number o equally spaced points
}
  \item{move}{
move the outline
}
  \item{move.factor}{
how much to move or translate the outline
}
  \item{plot.it}{
plot the outline or not
}
  \item{show.change}{
location the change is included
}
}
\details{
To generate one quadratic allow both linear spiral expansions to be equal (see example)
The spiral axis is very close to first outline point (umbo). 

Continuity conditions put severe limitations on the shape of the second quadratic.

This double quadratic is a NOT one of the standard spirals and has dubious value. McGhee (1980) has a 
quadratic formulation that may be better suited (see  curved logarithmic spiral and \code{fitQuadratic}).

}
\value{
 data frame with spiral coordinates as x, y, logr, angle, and arc length
}
\references{
Spiral notebook (page 10B) for proof and the exponential formulation.  

McGhee, G.R., 1980. Shell form in the biconvex articulate Brachiopoda: 
A geometric analysis.  Paleobiology 6: 57-76.
}
\author{
A.E. Aldridge
}
\note{
not generated by equal arc lengths.  
}
\seealso{
\code{generateAnyLinearSpirals}   	\code{     generateCurvedSpiral}
}
\examples{
\dontrun{
# ONE QUADRATIC SPIRAL using this function instead of generateCurvedSpiral
z <- generateQuadraticSpirals(k1=45,k2=45,q=0.03, dist.zero=2,theta.end=0.8,equal=F)
xsq <- fitOneQuadratic( x.df$x, x.df$y, axis.start=c(980,1000))
xsq$parameters1
xsq$parameters2
#
# TWO CONNECTED QUADRATIC SPIRALS ( if it makes sense or is even possible)
 z <- generateQuadraticSpirals(k1=45,k2=50, dist.zero=2,theta.end=0.8,equal=F)
xs <- fitTwoQuadratic( x=z$x, y=z$y, m=100, axis.start=c(980,1000))
## WARNING - THIS NOT WORKING WELL in the fitting (it is ok with generating)
## problem being that connection requires careful choice of input & start values

}
}
\keyword{ CurvedSpiral } 
